requirements.txt
yt_dlp
openai-whisper
google-generativeai
python-dotenv
rich
